Separate word-/part-based grammar checks into new file
